fix: re-export cockroach dialect and adapter separately from migrator
Addresses a problem where the migrator wouldn't run because the adapter wasn't exported, while still making it possible for nextjs to run without having to resolve the new URL in createMigrator
