+++ packages/ow-db/src/migrator.ts
@@ -4,24 +4,12 @@ import * as path from "path"
 
 import { Database } from "./types"
 
-export class CockroachAdapter extends PostgresAdapter {
-  override acquireMigrationLock(): Promise<void> {
-    return Promise.resolve()
-  }
-}
-
-export class CockroachDialect extends PostgresDialect {
-  override createAdapter(): CockroachAdapter {
-    return new CockroachAdapter()
-  }
-}
-
 export const createMigrator = (db: Kysely<Database>) =>
   new Migrator({
     db: db,
     provider: new FileMigrationProvider({
       fs,
       path,
-      migrationFolder: new URL("../src/migrations", import.meta.url).pathname,
+      migrationFolder: new URL("migrations", import.meta.url).pathname,
     }),
   })
